Fix C-style casts of size_t to int in EXPECT/ASSERT.
Follow-on for old-style casting from
  https://chromium-review.googlesource.com/c/chromium/src/+/7218814

Using unsigned literals avoids the need for such casts, and reduces the
number of potentially unsafe casts to be investigated down the road.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7226233
Owners-Override: Nico Weber <thakis@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1554618}

+++ chrome/browser/ash/arc/input_overlay/touch_injector_unittest.cc
@@ -522,7 +522,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
       kValidJsonActionTapKey, base::JSON_PARSE_CHROMIUM_EXTENSIONS);
   injector_->ParseActions(json_value->GetDict());
   // Extra Action with the same ID is removed.
-  EXPECT_EQ(2, (int)injector_->actions().size());
+  EXPECT_EQ(2u, injector_->actions().size());
   auto* actionA = injector_->actions()[0].get();
   auto* actionB = injector_->actions()[1].get();
   injector_->RegisterEventRewriter();
@@ -531,7 +531,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
   event_generator_->PressKey(ui::VKEY_A, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(actionA->touch_id());
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(1, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(1u, event_capturer_.touch_events().size());
   auto* event = event_capturer_.touch_events()[0].get();
   EXPECT_EQ(ui::EventType::kTouchPressed, event->type());
   auto expectA1 =
@@ -542,7 +542,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
   event_generator_->ReleaseKey(ui::VKEY_A, ui::EF_NONE, kTolerance);
   EXPECT_FALSE(actionA->touch_id());
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(2, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(2u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[1].get();
   EXPECT_EQ(ui::EventType::kTouchReleased, event->type());
   EXPECT_POINTF_NEAR(expectA1, event->root_location_f(), kTolerance);
@@ -552,23 +552,23 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
   // Unregister the event rewriter to see if extra events are sent.
   injector_->UnRegisterEventRewriter();
 
-  EXPECT_EQ(2, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(2u, event_capturer_.touch_events().size());
   event_capturer_.Clear();
   injector_->RegisterEventRewriter();
 
   // Press and release key C, it should receive key event, not touch event.
   event_generator_->PressKey(ui::VKEY_C, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.touch_events().empty());
-  EXPECT_EQ(1, (int)event_capturer_.key_events().size());
+  EXPECT_EQ(1u, event_capturer_.key_events().size());
   event_generator_->ReleaseKey(ui::VKEY_C, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.touch_events().empty());
-  EXPECT_EQ(2, (int)event_capturer_.key_events().size());
+  EXPECT_EQ(2u, event_capturer_.key_events().size());
   event_capturer_.Clear();
 
   // Test multi-key tap: Press B -> Press A -> Release A -> Release B.
   event_generator_->PressKey(ui::VKEY_B, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(1, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(1u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[0].get();
   EXPECT_EQ(ui::EventType::kTouchPressed, event->type());
   auto expectB =
@@ -578,7 +578,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
 
   event_generator_->PressKey(ui::VKEY_A, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(2, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(2u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[1].get();
   EXPECT_EQ(ui::EventType::kTouchPressed, event->type());
   auto expectA2 =
@@ -588,7 +588,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
 
   event_generator_->ReleaseKey(ui::VKEY_A, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(3, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(3u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[2].get();
   EXPECT_EQ(ui::EventType::kTouchReleased, event->type());
   EXPECT_POINTF_NEAR(expectA2, event->root_location_f(), kTolerance);
@@ -596,7 +596,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
 
   event_generator_->ReleaseKey(ui::VKEY_B, ui::EF_NONE, kTolerance);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(4, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(4u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[3].get();
   EXPECT_EQ(ui::EventType::kTouchReleased, event->type());
   EXPECT_POINTF_NEAR(expectB, event->root_location_f(), kTolerance);
@@ -606,7 +606,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
   // Test multi-key tap: Press B -> Press A -> Release B -> Release A.
   event_generator_->PressKey(ui::VKEY_B, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(1, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(1u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[0].get();
   EXPECT_EQ(ui::EventType::kTouchPressed, event->type());
   EXPECT_POINTF_NEAR(expectB, event->root_location_f(), kTolerance);
@@ -614,7 +614,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
 
   event_generator_->PressKey(ui::VKEY_A, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(2, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(2u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[1].get();
   EXPECT_EQ(ui::EventType::kTouchPressed, event->type());
   EXPECT_POINTF_NEAR(expectA1, event->root_location_f(), kTolerance);
@@ -622,7 +622,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
 
   event_generator_->ReleaseKey(ui::VKEY_B, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(3, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(3u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[2].get();
   EXPECT_EQ(ui::EventType::kTouchReleased, event->type());
   EXPECT_POINTF_NEAR(expectB, event->root_location_f(), kTolerance);
@@ -630,7 +630,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
 
   event_generator_->ReleaseKey(ui::VKEY_A, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(event_capturer_.key_events().empty());
-  EXPECT_EQ(4, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(4u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[3].get();
   EXPECT_EQ(ui::EventType::kTouchReleased, event->type());
   EXPECT_POINTF_NEAR(expectA1, event->root_location_f(), kTolerance);
@@ -641,7 +641,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
   event_generator_->PressKey(ui::VKEY_B, ui::EF_NONE, /*source_device_id=*/1);
   event_generator_->PressKey(ui::VKEY_B, ui::EF_IS_REPEAT,
                              /*source_device_id=*/1);
-  EXPECT_EQ(1, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(1u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[0].get();
   EXPECT_POINTF_NEAR(expectB, event->root_location_f(), kTolerance);
   event_generator_->ReleaseKey(ui::VKEY_B, ui::EF_NONE, /*source_device_id=*/1);
@@ -652,7 +652,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
   EXPECT_TRUE(actionB->touch_id());
   injector_->UnRegisterEventRewriter();
   EXPECT_FALSE(actionB->touch_id());
-  EXPECT_EQ(2, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(2u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events().back().get();
   EXPECT_EQ(ui::EventType::kTouchReleased, event->type());
   EXPECT_POINTF_NEAR(expectB, event->root_location_f(), kTolerance);
@@ -662,7 +662,7 @@ TEST_F(TouchInjectorTest, TestEventRewriterActionTapKey) {
   injector_->RegisterEventRewriter();
   event_generator_->PressKey(ui::VKEY_B, ui::EF_NONE, /*source_device_id=*/1);
   EXPECT_TRUE(actionB->touch_id());
-  EXPECT_EQ(1, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(1u, event_capturer_.touch_events().size());
   event = event_capturer_.touch_events()[0].get();
   EXPECT_EQ(ui::EventType::kTouchPressed, event->type());
   EXPECT_POINTF_NEAR(expectB, event->root_location_f(), kTolerance);
@@ -1230,7 +1230,7 @@ TEST_F(TouchInjectorTest, TestPlayWithGameControlsHistogramHistogramsYes) {
   auto json_value = base::JSONReader::ReadAndReturnValueWithError(
       kValidJsonActionTapKey, base::JSON_PARSE_CHROMIUM_EXTENSIONS);
   injector_->ParseActions(json_value->GetDict());
-  EXPECT_EQ(2, (int)injector_->actions().size());
+  EXPECT_EQ(2u, injector_->actions().size());
   injector_->RegisterEventRewriter();
 
   base::HistogramTester histograms;
@@ -1246,7 +1246,7 @@ TEST_F(TouchInjectorTest, TestPlayWithGameControlsHistogramHistogramsYes) {
   // Press and release key A, it plays with Game Controls.
   event_generator_->PressAndReleaseKey(ui::VKEY_A, ui::EF_NONE,
                                        /*source_device_id=*/1);
-  EXPECT_EQ(2, (int)event_capturer_.touch_events().size());
+  EXPECT_EQ(2u, event_capturer_.touch_events().size());
   VerifyPlayWithGameControlsHistogram(
       histograms, std::vector<int>{0, /*played_with_game_controls==true*/ 1});
   VerifyPlayGameWithGameControlsUkmEvent(
@@ -1268,7 +1268,7 @@ TEST_F(TouchInjectorTest, TestPlayWithGameControlsHistogramHistogramsNo) {
   auto json_value = base::JSONReader::ReadAndReturnValueWithError(
       kValidJsonActionTapKey, base::JSON_PARSE_CHROMIUM_EXTENSIONS);
   injector_->ParseActions(json_value->GetDict());
-  EXPECT_EQ(2, (int)injector_->actions().size());
+  EXPECT_EQ(2u, injector_->actions().size());
   injector_->RegisterEventRewriter();
 
   base::HistogramTester histograms;
