+++ components/commerce/core/shopping_service_unittest.cc
@@ -1677,7 +1677,7 @@ TEST_P(ShoppingServiceTest, TestPriceInsightsInfoResponse) {
             ASSERT_EQ(kLowTypicalPrice, info->typical_low_price_micros);
             ASSERT_EQ(kHighTypicalPrice, info->typical_high_price_micros);
             ASSERT_EQ(kAttributes, info->catalog_attributes);
-            ASSERT_EQ(2, (int)(info->catalog_history_prices.size()));
+            ASSERT_EQ(2u, info->catalog_history_prices.size());
             ASSERT_EQ("2021-01-01",
                       std::get<0>(info->catalog_history_prices[0]));
             ASSERT_EQ("2021-01-02",
@@ -1725,7 +1725,7 @@ TEST_P(ShoppingServiceTest,
             ASSERT_EQ(kLowTypicalPrice, info->typical_low_price_micros);
             ASSERT_EQ(kHighTypicalPrice, info->typical_high_price_micros);
             ASSERT_EQ(std::nullopt, info->catalog_attributes);
-            ASSERT_EQ(0, (int)(info->catalog_history_prices.size()));
+            ASSERT_EQ(0u, info->catalog_history_prices.size());
             ASSERT_EQ(std::nullopt, info->jackpot_url);
             ASSERT_EQ(PriceBucket::kHighPrice, info->price_bucket);
             ASSERT_EQ(true, info->has_multiple_catalogs);
@@ -1796,7 +1796,7 @@ TEST_P(ShoppingServiceTest, TestPriceInsightsInfoResponse_EmptyRange) {
             ASSERT_EQ(std::nullopt, info->typical_low_price_micros);
             ASSERT_EQ(std::nullopt, info->typical_high_price_micros);
             ASSERT_EQ(kAttributes, info->catalog_attributes);
-            ASSERT_EQ(2, (int)(info->catalog_history_prices.size()));
+            ASSERT_EQ(2u, info->catalog_history_prices.size());
             ASSERT_EQ("2021-01-01",
                       std::get<0>(info->catalog_history_prices[0]));
             ASSERT_EQ("2021-01-02",
@@ -2084,7 +2084,7 @@ TEST_P(ShoppingServiceTest, TestDiscountInfoResponse_ForMerchant) {
       GURL(kDiscountsUrl1),
       base::BindOnce([](const GURL& url,
                         const std::vector<DiscountInfo> discounts) {
-        ASSERT_EQ(1, (int)discounts.size());
+        ASSERT_EQ(1u, discounts.size());
         ASSERT_TRUE(discounts[0].expiry_time_sec.has_value());
       }).Then(run_loop.QuitClosure()));
   run_loop.Run();
@@ -2123,7 +2123,7 @@ TEST_P(ShoppingServiceTest, TestDiscountInfoResponse) {
       base::BindOnce(
           [](base::RunLoop* run_loop, const GURL& url,
              const std::vector<DiscountInfo> discounts) {
-            ASSERT_EQ(1, (int)discounts.size());
+            ASSERT_EQ(1u, discounts.size());
 
             ASSERT_EQ(DiscountClusterType::kOfferLevel,
                       discounts[0].cluster_type);
@@ -2177,7 +2177,7 @@ TEST_P(ShoppingServiceTest,
       GURL(kDiscountsUrl1),
       base::BindOnce([](const GURL& url,
                         const std::vector<DiscountInfo> discounts) {
-        ASSERT_EQ(1, (int)discounts.size());
+        ASSERT_EQ(1u, discounts.size());
         ASSERT_EQ(DiscountClusterType::kOfferLevel, discounts[0].cluster_type);
         ASSERT_EQ(DiscountType::kCrawledPromotion, discounts[0].type);
         ASSERT_FALSE(discounts[0].expiry_time_sec.has_value());
@@ -2221,7 +2221,7 @@ TEST_P(ShoppingServiceTest, TestDiscountInfoResponse_InfoWithoutId) {
       GURL(kDiscountsUrl1), base::BindOnce(
                                 [](base::RunLoop* run_loop, const GURL& url,
                                    const std::vector<DiscountInfo> discounts) {
-                                  ASSERT_EQ(1, (int)discounts.size());
+                                  ASSERT_EQ(1u, discounts.size());
                                   ASSERT_EQ(kDiscountId1, discounts[0].id);
                                   run_loop->Quit();
                                 },
@@ -2262,7 +2262,7 @@ TEST_P(ShoppingServiceTest, TestDiscountInfoResponse_InfoWithoutTerms) {
       base::BindOnce(
           [](base::RunLoop* run_loop, const GURL& key,
              const std::vector<DiscountInfo> discounts) {
-            ASSERT_EQ(1, (int)discounts.size());
+            ASSERT_EQ(1u, discounts.size());
             ASSERT_EQ(kDiscountId1, discounts[0].id);
             ASSERT_FALSE(discounts[0].terms_and_conditions.has_value());
             run_loop->Quit();
@@ -2301,7 +2301,7 @@ TEST_P(ShoppingServiceTest, TestDiscountInfoResponse_InfoWithoutDiscountCode) {
       GURL(kDiscountsUrl1), base::BindOnce(
                                 [](base::RunLoop* run_loop, const GURL& key,
                                    const std::vector<DiscountInfo> discounts) {
-                                  ASSERT_EQ(0, (int)discounts.size());
+                                  ASSERT_EQ(0u, discounts.size());
                                   run_loop->Quit();
                                 },
                                 &run_loop));
@@ -2338,7 +2338,7 @@ TEST_P(ShoppingServiceTest, TestDiscountInfoResponse_InfoWithUnspecifiedType) {
       GURL(kDiscountsUrl1), base::BindOnce(
                                 [](base::RunLoop* run_loop, const GURL& key,
                                    const std::vector<DiscountInfo> discounts) {
-                                  ASSERT_EQ(0, (int)discounts.size());
+                                  ASSERT_EQ(0u, discounts.size());
                                   run_loop->Quit();
                                 },
                                 &run_loop));
